Refactor background handling
