+++ CustomTracks/CustomTrack.cs
@@ -4,11 +4,8 @@ using System.IO;
 using System.Linq;
 using BaboonAPI.Hooks.Tracks;
 using Newtonsoft.Json;
-using TrombLoader.Data;
-using HarmonyLib;
-using TrombLoader.Helpers;
+using TrombLoader.CustomTracks.Backgrounds;
 using UnityEngine;
-using UnityEngine.Video;
 
 namespace TrombLoader.CustomTracks;
 
@@ -108,7 +105,31 @@ public class CustomTrack : TromboneTrack
 
     public LoadedTromboneTrack LoadTrack()
     {
-        return new LoadedCustomTrack(this);
+        return new LoadedCustomTrack(this, LoadBackground());
+    }
+
+    private AbstractBackground LoadBackground()
+    {
+        if (File.Exists(Path.Combine(folderPath, "bg.trombackground")))
+        {
+            var bundle = AssetBundle.LoadFromFile(Path.Combine(folderPath, "bg.trombackground"));
+            return new CustomBackground(bundle, folderPath);
+        }
+        
+        var possibleVideoPath = Path.Combine(folderPath, "bg.mp4");
+        if (File.Exists(possibleVideoPath))
+        {
+            return new VideoBackground(possibleVideoPath);
+        }
+
+        var spritePath = Path.Combine(folderPath, "bg.png");
+        if (File.Exists(spritePath))
+        {
+            return new ImageBackground(spritePath);
+        }
+
+        Plugin.LogWarning($"No background for track {trackref}");
+        return new EmptyBackground();
     }
 
     public bool IsVisible()
@@ -116,26 +137,15 @@ public class CustomTrack : TromboneTrack
         return true;
     }
 
-    public enum BackgroundType
-    {
-        Image,
-        Video,
-        Trombackground,
-        Empty
-    }
-
     public class LoadedCustomTrack : LoadedTromboneTrack
     {
-        private CustomTrack _parent;
-        private AssetBundle _backgroundBundle;
+        private readonly CustomTrack _parent;
+        private readonly AbstractBackground _background;
 
-        private string videoPath = null;
-        private string imagePath = null;
-        private BackgroundType _backgroundType = BackgroundType.Empty;
-
-        public LoadedCustomTrack(CustomTrack parent)
+        public LoadedCustomTrack(CustomTrack parent, AbstractBackground background)
         {
             _parent = parent;
+            _background = background;
         }
 
         public TrackAudio LoadAudio()
@@ -162,85 +172,7 @@ public class CustomTrack : TromboneTrack
 
         public GameObject LoadBackground(BackgroundContext ctx)
         {
-            var songPath = _parent.folderPath;
-            if (File.Exists(Path.Combine(songPath, "bg.trombackground")))
-            {
-                _backgroundType = BackgroundType.Trombackground;
-                _backgroundBundle = AssetBundle.LoadFromFile(Path.Combine(songPath, "bg.trombackground"));
-                var bg = _backgroundBundle.LoadAsset<GameObject>("assets/_background.prefab");
-
-                // MacOS Shader Handling
-                // May need to be expanded to other platforms eventually, but for now only check for invalid shaders on mac
-                // TODO: Handle custom shaders and the bundle that TrombLoaderBackgroundProject will eventually export
-                if (Application.platform == RuntimePlatform.OSXPlayer)
-                {
-                    foreach (var renderer in bg.GetComponentsInChildren<Renderer>(true))
-                    {
-                        foreach (var material in renderer.sharedMaterials)
-                        {
-                            if (material == null || material.shader == null || material.shader.isSupported) continue;
-
-                            Shader shader;
-                            if (Plugin.Instance.ShaderAssets.TryGetValue(material.shader.name, out shader))
-                            {
-                                // Shader exists and is cached, so *hopefully* it's the same and can be swapped out with no ill effects.
-                                material.shader = shader;
-                            }
-                            else
-                            {
-                                // TODO: Handle more gracefully. Maybe replace with a default shader.
-                                Plugin.LogDebug($"Could not find shader on {renderer.gameObject.name} ({material.shader.name})");
-                            }
-                        }
-                    }
-                }
-
-                var managers = bg.GetComponentsInChildren<TromboneEventManager>();
-                foreach (var eventManager in managers)
-                {
-                    eventManager.DeserializeAllGenericEvents();
-                }
-
-                var invoker = bg.AddComponent<TromboneEventInvoker>();
-                invoker.InitializeInvoker(ctx.controller, managers);
-
-                foreach (var videoPlayer in bg.GetComponentsInChildren<VideoPlayer>())
-                {
-                    if (videoPlayer.url == null || !videoPlayer.url.Contains("SERIALIZED_OUTSIDE_BUNDLE")) continue;
-                    var videoName = videoPlayer.url.Replace("SERIALIZED_OUTSIDE_BUNDLE/", "");
-                    var clipURL = Path.Combine(songPath, videoName);
-                    videoPlayer.url = clipURL;
-                }
-
-                BackgroundHelper.ApplyBackgroundTransforms(ctx.controller, bg);
-
-                return bg;
-            }
-            else
-            {
-                _backgroundBundle = AssetBundle.LoadFromFile($"{Application.streamingAssetsPath}/trackassets/ballgame");
-                var bg = _backgroundBundle.LoadAsset<GameObject>("BGCam_ballgame");
-
-                var possibleVideoPath = Path.Combine(songPath, "bg.mp4");
-                if (File.Exists(possibleVideoPath))
-                {
-                    _backgroundType = BackgroundType.Video;
-                    videoPath = possibleVideoPath;
-                    return bg;
-                }
-
-                var spritePath = Path.Combine(songPath, "bg.png");
-                if (File.Exists(spritePath))
-                {
-                    _backgroundType = BackgroundType.Image;
-                    imagePath = spritePath;
-                    return bg;
-                }
-
-                Plugin.LogWarning($"No background for track {_parent.trackref}");
-                _backgroundType = BackgroundType.Empty;
-                return bg;
-            }
+            return _background.Load(ctx);
         }
 
         public void SetUpBackgroundDelayed(BGController controller, GameObject bg)
@@ -253,28 +185,7 @@ public class CustomTrack : TromboneTrack
                 modelCam.clearFlags = CameraClearFlags.Depth;
             }
 
-            switch (_backgroundType)
-            {
-                case BackgroundType.Image:
-                    BackgroundHelper.ApplyImage(bg, imagePath);
-                    break;
-                case BackgroundType.Video:
-                    BackgroundHelper.ApplyVideo(bg, controller, videoPath);
-                    break;
-                case BackgroundType.Trombackground:
-                    // Set up tromboners for trombackgrounds
-                    BackgroundHelper.SetUpPuppets(controller, bg);
-
-                    // Preload videos
-                    foreach (var videoPlayer in bg.GetComponentsInChildren<VideoPlayer>())
-                    {
-                        videoPlayer.Prepare();
-                    }
-                    break;
-                case BackgroundType.Empty:
-                    BackgroundHelper.DisableBackground(bg);
-                    break;
-            }
+            _background.SetUpBackground(controller, bg);
 
             controller.tickontempo = false;
 
@@ -284,10 +195,7 @@ public class CustomTrack : TromboneTrack
 
         public void Dispose()
         {
-            if (_backgroundBundle != null)
-            {
-                _backgroundBundle.Unload(false);
-            }
+            _background.Dispose();
         }
 
         public string trackref => _parent.trackref;
